pyproject.toml
[project]
name = "cl-ai-chat"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "colorama>=0.4.6",
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-ollama>=1.0.1",
    "langgraph>=1.0.9",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
]
